Added 7 quotes on spanish (#3953) dioswilson
* Updated spanish.json with 11 quotes

Added 11 quotes in spanish

* Fixed typos

Fixed typos I made on the previous quotes I made on spanish

* Last typo i found on the quotes

* Added more 7 quotes to spanish

Added 7 quotes to spanish and deleted one word that was wrong from a prevoius quote

* Changed quotation marks

One quote uses quotation marks, now it is as \", not sure if that's the correct way

* Updated quote's lenght which was wrong

* Changed swear from last quote.

Changed the swear on quote 126, new translation:
This "New Era" that everyone is talking about is trash. the time where pirates dreams have ended?! People dreams never end!