fix(language): fixed Marathi not showing up in the language list (@agokule) (#6395)
### Description

Right now, the Marathi language should exist on Monkeytype (there is a
`marathi.json` file in both `frontend/static/quotes` and
`frontend/static/languages`) but it doesn't. This is probably because it
isn't in the `_list.json` and `_groups.json` file, so I added it to
those files.

This is related to #4174.

---------

Co-authored-by: Atharv Gokule <106445077+username121546434@users.noreply.github.com>